pyproject.toml
[project]
name = "domainidom"
version = "0.1.0"
description = "Domain and brand name ideation, availability research, and packaging"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name="HamHockHQ", email="howardmd1@gmail.com"}]
dependencies = [
  "fastapi~=0.115",
  "uvicorn[standard]~=0.30",
  "typer[all]~=0.12",
  "pydantic~=2.8",
  "httpx~=0.27",
  "python-dotenv~=1.0",
  "metaphone~=0.6",
  "python-Levenshtein~=0.25",
]

[project.optional-dependencies]
dev = [
  "pytest~=8.3",
  "black~=24.10",
  "ruff~=0.6",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
exclude = [".venv", "dist", "build"]

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["."]
